Extra js file for leaflet Map
