+++ app/javascript/packs/leafletMap.js
@@ -1,0 +1,92 @@
+let map;
+let routeLayer;
+
+export async function setupMap() {
+    map = L.map("map");
+    L.tileLayer("http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png", {
+        attribution:
+            '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a>',
+        maxZoom: 19,
+    }).addTo(map);
+
+    // these operations are executed in parallel
+    getView().then((view) => {
+        map.setView(view["latlng"], view["zoom"]);
+    });
+
+    getBuildings().then((buildingPolygons) => {
+        addPolygons(buildingPolygons);
+    });
+
+    getBuildingMarkers().then((buildingMarkers) => {
+        addMarkers(buildingMarkers);
+    });
+}
+
+export function addPolygons(polygons) {
+    polygons.forEach((polygon) => {
+        L.polygon(polygon["latlngs"], polygon["options"]).addTo(map);
+    });
+}
+
+export function addMarkers(markers, layer = map) {
+    markers.forEach((marker) => {
+        addMarker(marker, layer);
+    });
+}
+
+export function addMarker(marker, layer = map) {
+    marker["divIcon"]["iconSize"] = [];
+    marker["divIcon"]["iconAnchor"] = [0, 0];
+    marker["divIcon"]["popupAnchor"] = [0, 0];
+    const icon = L.divIcon(marker["divIcon"]);
+    L.marker(marker["latlng"], { icon: icon }).addTo(layer);
+}
+
+export function addPolylines(polylines, layer = map) {
+    polylines.forEach((polyline) => {
+        addPolyline(polyline, layer);
+    });
+}
+
+export function addPolyline(polyline, layer = map) {
+    L.polyline(polyline["latlngs"], polyline["options"]).addTo(layer);
+}
+
+export async function displayRoute(start, dest) {
+    const route = await $.ajax({
+        type: "GET",
+        url: "building_map_route",
+        data: `start=${start}&dest=${dest}`,
+        dataType: "json",
+    });
+    if (routeLayer) routeLayer.clearLayers();
+    else routeLayer = L.layerGroup();
+    addPolyline(route["polyline"], routeLayer);
+    addMarkers([route["marker"]], routeLayer);
+    routeLayer.addTo(map);
+}
+
+async function getBuildings() {
+    return $.ajax({
+        type: "GET",
+        url: "building_map_buildings",
+        dataType: "json",
+    });
+}
+
+async function getBuildingMarkers() {
+    return $.ajax({
+        type: "GET",
+        url: "building_map_markers",
+        dataType: "json",
+    });
+}
+
+async function getView() {
+    return $.ajax({
+        type: "GET",
+        url: "building_map_view",
+        dataType: "json",
+    });
+}
