+++ app/javascript/packs/locations.js
@@ -1,0 +1,3 @@
+import { setupMap } from "./leafletMap.js";
+
+setupMap();
