+++ app/views/building_map/index.html.erb
@@ -62,45 +62,3 @@
 </div>
 
 <%= javascript_pack_tag 'building_map' %>
-
-<%# Indoor labels %>
-<script>
-  function addIndoorLabel(feature, layer) {
-    layer.bindTooltip(
-      feature.properties.name,
-      {
-          permanent: true,
-          direction:'center',
-          className: 'indoor-label',
-      }
-    );
-  }
-
-  function loadGeoJsonFile(filename) {
-    fetch(filename)
-    .then(response => response.json())
-    .then(geojsonFeatureCollection => {
-      // Manually add indoor labels to map
-      const rooms = L.geoJSON(geojsonFeatureCollection, { onEachFeature: addIndoorLabel }).addTo(map);
-      rooms.eachLayer((layer)=>{
-        layer.getTooltip().setLatLng(layer.getBounds().getCenter());
-      });
-      recalculateTooltipVisibility()
-    })
-  }
-
-  loadGeoJsonFile('assets/lecture-hall-building.geojson')
-  function recalculateTooltipVisibility() {
-    const zoomLevel = map.getZoom()
-    map.eachLayer(layer => {
-      if (layer.getTooltip()) {
-        if (zoomLevel == 19 /* nearest zoom */) {
-          layer.openTooltip(layer.getBounds().getCenter())
-        } else {
-          layer.closeTooltip()
-        }
-      }
-    })
-  }
-  map.on('zoomend', recalculateTooltipVisibility);
-</script>
