+++ app/views/locations/show.html.erb
@@ -34,7 +34,7 @@
 </div>
 
 <%# TODO: This passes all the tests altough for now no markers are added to the map. %>
-<%= javascript_pack_tag 'building_map' %>
+<%= javascript_pack_tag 'leafletMap' %>
 
 <p>
   <strong>Opening Times:</strong>
